complete mapping modularization refactor
